удалил большую часть eslint заглушек, остальные или пока не получиться удалить, или надо добавлять исключения
