+++ src/common/errorBlock/ErrorBlock.tsx
@@ -1,4 +1,3 @@
-/* eslint-disable @typescript-eslint/no-use-before-define */
 import * as React from 'react';
 import StyledError from './styles';
