+++ src/common/tags/Tag.tsx
@@ -1,4 +1,3 @@
-/* eslint-disable @typescript-eslint/no-use-before-define */
 import React from 'react';
 
 import { StyledTag, TagText, CloseButton, ClosePicture } from './style';
