+++ src/components/Audio/AudioPage.tsx
@@ -85,7 +85,6 @@ interface IBtnFilterAudio {
 }
 
 interface ISlickOnClick {
-  // eslint-disable-next-line react/require-default-props
   onClick?: () => void;
 }
