+++ src/components/Group/ModalEdit.tsx
@@ -1,5 +1,3 @@
-/* eslint-disable no-shadow */
-/* eslint-disable jsx-a11y/label-has-associated-control */
 import React from 'react';
 import styled from 'styled-components';
 import { useForm } from 'react-hook-form';
