--- src/components/Main/UserAbout/models.ts
@@ -1,12 +1,0 @@
-/* eslint-disable import/prefer-default-export */
-const desc = {
-  birthday: 'День рождения',
-  educ: 'Образование',
-  prof: 'Профессия',
-  lang: 'Язык',
-  city: 'Город',
-  about: 'Личная информация',
-  site: 'Личный сайт',
-};
-
-export default desc;
