+++ src/redux-toolkit/audios/allAudiosSlice.ts
@@ -1,4 +1,3 @@
-/* eslint-disable max-len */
 import { createAsyncThunk, createSlice, Draft, PayloadAction } from '@reduxjs/toolkit';
 import {
   fetchAudiosAll,
